requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
joblib==1.3.2
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.2
oauthlib==3.2.2
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
